# openapi/openapi.yaml
openapi: 3.1.0
info:
  title: OpenRating API
  version: 1.0.0
servers:
  - url: https://api.example.com/v1
paths:
...
  /v1/events/{event_id}/competitions:
    post:
      tags: [Competitions]
      summary: Create a competition within an event
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionCreate'
      responses:
        '201':
          description: Competition created
          content:
            application/json:
              schema:
                type: object
                properties:
                  competition:
                    $ref: '#/components/schemas/Competition'
        '400':
          description: Invalid request
        '404':
          description: Event not found
    get:
      tags: [Competitions]
      summary: List competitions for an event
      parameters:
        - in: path
          name: event_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Competitions for the event
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/CompetitionListResponse'
        '404':
          description: Event not found
  /v1/competitions/{competition_id}:
    get:
      tags: [Competitions]
      summary: Retrieve a competition
      parameters:
        - in: path
          name: competition_id
          required: true
          schema: { type: string }
      responses:
        '200':
          description: Competition details
          content:
            application/json:
              schema:
                type: object
                properties:
                  competition:
                    $ref: '#/components/schemas/Competition'
        '404':
          description: Competition not found
    patch:
      tags: [Competitions]
      summary: Update a competition
      parameters:
        - in: path
          name: competition_id
          required: true
          schema: { type: string }
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/CompetitionUpdate'
      responses:
        '200':
          description: Updated competition
          content:
            application/json:
              schema:
                type: object
                properties:
                  competition:
                    $ref: '#/components/schemas/Competition'
        '400':
          description: Invalid request
        '404':
          description: Competition not found
components:
  schemas:
...
    Competition:
      type: object
      properties:
        competition_id: { type: string }
        event_id: { type: string }
        event_slug: { type: string }
        organization_id: { type: string }
        name: { type: string }
        slug: { type: string }
        sport:
          type: [string, 'null']
          enum: [BADMINTON, TENNIS, SQUASH, PADEL, PICKLEBALL]
        discipline:
          type: [string, 'null']
          enum: [SINGLES, DOUBLES, MIXED]
        format:
          type: [string, 'null']
        tier:
          type: [string, 'null']
        status:
          type: [string, 'null']
        draw_size:
          type: [integer, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        metadata:
          type: [object, 'null']
          additionalProperties: true
        created_at:
          type: [string, 'null']
          format: date-time
        updated_at:
          type: [string, 'null']
          format: date-time
    CompetitionCreate:
      type: object
      required: [name]
      properties:
        name: { type: string }
        slug: { type: string }
        sport:
          type: [string, 'null']
          enum: [BADMINTON, TENNIS, SQUASH, PADEL, PICKLEBALL]
        discipline:
          type: [string, 'null']
          enum: [SINGLES, DOUBLES, MIXED]
        format:
          type: [string, 'null']
        tier:
          type: [string, 'null']
        status:
          type: [string, 'null']
        draw_size:
          type: [integer, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        metadata:
          type: [object, 'null']
          additionalProperties: true
    CompetitionUpdate:
      type: object
      properties:
        name: { type: string }
        slug: { type: string }
        sport:
          type: [string, 'null']
          enum: [BADMINTON, TENNIS, SQUASH, PADEL, PICKLEBALL]
        discipline:
          type: [string, 'null']
          enum: [SINGLES, DOUBLES, MIXED]
        format:
          type: [string, 'null']
        tier:
          type: [string, 'null']
        status:
          type: [string, 'null']
        draw_size:
          type: [integer, 'null']
        start_date:
          type: [string, 'null']
          format: date-time
        end_date:
          type: [string, 'null']
          format: date-time
        metadata:
          type: [object, 'null']
          additionalProperties: true
    CompetitionListResponse:
      type: object
      properties:
        competitions:
          type: array
          items:
            $ref: '#/components/schemas/Competition'
...